+++ src/by/tms/action/util/Reader.java
@@ -2,6 +2,8 @@ package by.tms.action.util;
 
 import java.util.Scanner;
 
+import static by.tms.action.util.Writer.writeln;
+
 public class Reader {
     public static double readDouble() { // метод использовать там, где нужны только цифры
         Scanner scanner = new Scanner(System.in);
@@ -40,5 +42,21 @@ public class Reader {
         }
         return null;
     }
-}
 
+    public static String readWithInvite(String invite) {
+        Writer.write(invite);
+        return readLine();
+    }
+
+    public static long readId() {
+        while (true) {
+            try {
+                long id = Long.parseLong(readWithInvite("Input Student ID: "));
+                if (id < 0) throw new NumberFormatException();
+                return id;
+            } catch (NumberFormatException e) {
+                writeln("Id incorrect! Try again!");
+            }
+        }
+    }
+}
