+++ src/by/tms/action/util/Writer.java
@@ -1,8 +1,12 @@
 package by.tms.action.util;
 
 public class Writer {
+
     public static void write(Object message) {
-        System.out.println(message);
+        System.out.print(message);
     }
 
+    public static void writeln(Object message) {
+        System.out.println(message);
+    }
 }
