--- src/by/tms/util/Writer.java
@@ -1,12 +1,0 @@
-package by.tms.util;
-
-public class Writer {
-
-    public static void write(Object message) {
-        System.out.print(message);
-    }
-
-    public static void writeln(Object message) {
-        System.out.println(message);
-    }
-}
